Advisory Bitmap
Replace raw stdint/stdbool usage with unified Core/Fundtypes types (BYTE/WORD/DWORD/QWORD/USIZE/UINTPTR/BOOL/NULLPTR) across CPU, KRTL and memory headers/sources and update function signatures accordingly. Add virtual memory management API (Kernel/Include/Memory/Virtmemmgmt.h) with stub implementation. Rework physical memory manager: introduce advisory vs dynamic bitmaps, compute page counts from discovered highest address, add KrRelinquishPhysicalPage/KrReservePhysicalPage/KrIsPhysicalPageReserved APIs, fix bitmap sizing/limits, and copy advisory bitmap to dynamic bitmap at init. Add runtime PMM test and reporting in KrKernelStart, plus BSS zeroing using new linker symbols (__KR_LINK_BSS_START/END) added to Kernel.ld. Improve Krnlmeltdown formatting and behavior (set bMeltdown, safer halt loop), add nicer Halt diagnostics that respect display protocol state, add new display color and active flag to DisplaywideTextProtocol, and several bugfixes (MSR rd/wr types, interrupt handler BOOL returns, BootstrapArena acquire guard for zero size, IDT/APIC type adjustments, minor formatting/assembly fixes). Also add new source file for Virtmemmgmt.c (stub) and various small refactors and API renames for clarity.

+++ Kernel/Include/CPU/MSR.h
@@ -1,11 +1,11 @@
 #ifndef YCH_KERNEL_CPU_MSR_H
 #define YCH_KERNEL_CPU_MSR_H
 
-#include <stdint.h>
+#include "Core/Fundtypes.h"
 
 // `RDMSR`
-uint64_t KrReadModelSpecificRegister(uint32_t idRegister);
+QWORD KrReadModelSpecificRegister(DWORD idRegister);
 // `WRMSR`
-void     KrWriteModelSpecificRegister(uint32_t idRegister, uint64_t data);
+void  KrWriteModelSpecificRegister(DWORD idRegister, DWORD data);
 
 #endif // !YCH_KERNEL_CPU_MSR_H
